Extract common S3 functionality from PutS3Object
